Revert "MU WPCOM: Prevent site owner from editing user's account-level fields (#42027)"
This reverts commit 71790f0c8d1906bf2918349acf41c13c7e53b4af.
