--- projects/packages/jetpack-mu-wpcom/changelog/feat-disable-account-fields-on-edit-user
@@ -1,4 +1,0 @@
-Significance: minor
-Type: changed
-
-MU WPCOM: Prevent site owner from editing user's account-level fields
