+++ projects/packages/jetpack-mu-wpcom/src/utils.php
@@ -154,30 +154,3 @@ function get_wpcom_blog_id() {
 
 	return false;
 }
-
-/**
- * Check if the site is a WordPress.com Atomic site.
- *
- * @return bool
- */
-function is_woa_site() {
-	if ( ! class_exists( 'Automattic\Jetpack\Status\Host' ) ) {
-		return false;
-	}
-	$host = new Automattic\Jetpack\Status\Host();
-	return $host->is_woa_site();
-}
-
-/**
- * Whether the current user is connected to WordPress.com.
- *
- * @param int $user_id the user identifier. Default is the current user.
- * @return bool Boolean is the user connected?
- */
-function is_user_connected( $user_id ) {
-	if ( defined( 'IS_WPCOM' ) && IS_WPCOM ) {
-		return true;
-	}
-
-	return ( new Connection_Manager( 'jetpack' ) )->is_user_connected( $user_id );
-}
